:hammer: Refactor episode details view classes directory
