+++ app/src/main/java/com/rummenigged/popcorntime/view/episodeDetails/EpisodeDetailsFragment.kt
@@ -1,4 +1,4 @@
-package com.rummenigged.popcorntime.view
+package com.rummenigged.popcorntime.view.episodeDetails
 
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.navArgs
